fix(evaluate): fractional rollout decode crash and stale client format version (#17)
RolloutSplit.Percentage was int, so any wire response containing a
sub-1% rollout split would fail JSON decoding outright and crash the
entire configuration fetch, not just that one flag's evaluation.
Changed to float64 and fixed bucketPositions() to multiply before
truncating.

clientFormatVersion was still hardcoded to 1 even though task 5.8 had
already implemented every FormatVersion2 construct (composed
ClauseTree, individual targets, prerequisites) here. The platform has
therefore been silently treating this client as FormatVersion1-only
in production, marking those constructs non_evaluable regardless of
this SDK's real capability. Bumped to 2.

Closes the task 9.2 fixture-coverage gap: added
TestTargetingModel_RolloutVectors covering the 2 in-scope rollout
vectors from the shared conformance fixture (segment vectors excluded
per task 7.7's structural finding — this SDK never needs
segment-handling code since the platform always resolves segment
membership server-side before serving a Configuration).

+++ rollout_outcome_test.go
@@ -39,7 +39,7 @@ func (v rolloutOutcomeVector) buildFlagConfig() rollfuse.FlagConfig {
 	splits := make([]rollfuse.RolloutSplit, 0, len(v.Rollout))
 
 	for _, s := range v.Rollout {
-		splits = append(splits, rollfuse.RolloutSplit{VariationKey: s.VariationKey, Percentage: s.Percentage})
+		splits = append(splits, rollfuse.RolloutSplit{VariationKey: s.VariationKey, Percentage: float64(s.Percentage)})
 
 		if !seen[s.VariationKey] {
 			seen[s.VariationKey] = true
